Move terminalEnvironment, node->common
Part of #69115
