+++ src/tsconfig.strictNullChecks.json
@@ -286,6 +286,7 @@
 		"./vs/workbench/contrib/terminal/common/terminal.ts",
 		"./vs/workbench/contrib/terminal/common/terminalColorRegistry.ts",
 		"./vs/workbench/contrib/terminal/common/terminalCommands.ts",
+		"./vs/workbench/contrib/terminal/common/terminalEnvironment.ts",
 		"./vs/workbench/contrib/terminal/common/terminalMenu.ts",
 		"./vs/workbench/contrib/terminal/common/terminalService.ts",
 		"./vs/workbench/contrib/terminal/electron-browser/terminalActions.ts",
@@ -295,7 +296,6 @@
 		"./vs/workbench/contrib/terminal/electron-browser/terminalProcessManager.ts",
 		"./vs/workbench/contrib/terminal/node/terminal.ts",
 		"./vs/workbench/contrib/terminal/node/terminalCommandTracker.ts",
-		"./vs/workbench/contrib/terminal/node/terminalEnvironment.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcess.ts",
 		"./vs/workbench/contrib/terminal/node/terminalProcessExtHostProxy.ts",
 		"./vs/workbench/contrib/terminal/node/windowsShellHelper.ts",
